feat: refactor Home component structure and styling, migrating to CSS Modules for improved maintainability and visual appeal
