+++ astronautmarkus/src/Components/Home/Home.module.css
@@ -1,0 +1,72 @@
+.headerImage img {
+    width: 250px;
+    height: auto;
+    animation: float 4s ease-in-out infinite;
+}
+
+@media (max-width: 768px) {
+    .headerImage img {
+        width: 200px;
+    }
+}
+
+@media (max-width: 480px) {
+    .headerImage img {
+        width: 150px;
+    }
+}
+
+.cta-buttons {
+    display: flex;
+    justify-content: center;
+    gap: 1rem;
+}
+
+@media (max-width: 480px) {
+    .cta-buttons {
+        flex-direction: column;
+        align-items: center;
+    }
+
+    .cta-buttons .btn {
+        width: 100%;
+        margin-bottom: 0.5rem;
+    }
+}
+
+.purpleButton {
+    background-color: #6a0dad;
+    color: white;
+    border: none;
+    padding: 0.75rem 1.5rem;
+    font-size: 1rem;
+    border-radius: 0.5rem;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+.purpleButton:hover {
+    background-color: #5a0c9d;
+}
+
+.purpleButton:focus {
+    outline: none;
+    box-shadow: 0 0 0 0.2rem rgba(106, 13, 173, 0.5);
+}
+
+@media (max-width: 480px) {
+    .purpleButton {
+        width: 100%;
+        margin-bottom: 0.5rem;
+    }
+}
+
+@keyframes float {
+    0%, 100% {
+        transform: translateY(0);
+    }
+    50% {
+        transform: translateY(-15px);
+    }
+}
+
