--- astronautmarkus/src/views/Home.module.css
@@ -1,16 +1,0 @@
-.headerImage img {
-    width: 250px;
-    height: auto;
-    animation: float 4s ease-in-out infinite;
-  }
-  
-  @keyframes float {
-    0%, 100% {
-      transform: translateY(0);
-    }
-    50% {
-      transform: translateY(-15px);
-    }
-  }
-  
-
